requirements.txt
django
mock
djangorestframework
markdown
django-filter
